Rename gammapy.spectrum to more consistent file naming scheme in Gammapy
