+++ gammapy/spectrum/extract.py
