+++ gammapy/spectrum/fit.py
