+++ gammapy/spectrum/group.py
