+++ gammapy/spectrum/simulation.py
