+++ gammapy/spectrum/tests/test_fit.py
