+++ gammapy/spectrum/tests/test_group.py
