+++ gammapy/spectrum/tests/test_observation.py
